+++ packages/vector_graphics/CHANGELOG.md
@@ -7,6 +7,7 @@
 * Adds SVG `feOffset` support with parameter, pixel, and visual reference tests.
 * Adds SVG `feColorMatrix` support with parameter, pixel, and visual reference tests.
 * Adds SVG `feFlood` support with parameter, pixel, and visual reference tests.
+* Adds SVG `feGaussianBlur` support with parameter, pixel, and visual reference tests.
 
 ## 1.2.3